% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_createLink.R
\name{utils_createLink}
\alias{utils_createLink}
\title{Change Domain to Hyperlink}
\usage{
utils_createLink(val1, val2)
}
\arguments{
\item{val1}{href domain}

\item{val2}{text}
}
\value{
string
}
\description{
Yup
}
