% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_createNumList.R
\name{utils_createNumList}
\alias{utils_createNumList}
\title{Create Numbered List}
\usage{
utils_createNumList(Names, nums = NA)
}
\arguments{
\item{Names}{vector of strings}
}
\value{
list
}
\description{
Yup
}
